Rename class file to its constructor name
