+++ packages/beacon-node/src/network/events.ts
@@ -2,7 +2,7 @@ import {EventEmitter} from "events";
 import {PeerId} from "@libp2p/interface-peer-id";
 import StrictEventEmitter from "strict-event-emitter-types";
 import {allForks, phase0} from "@lodestar/types";
-import {RequestTypedContainer} from "./reqresp/index.js";
+import {RequestTypedContainer} from "./reqresp/ReqRespBeaconNode.js";
 
 export enum NetworkEvent {
   /** A relevant peer has connected or has been re-STATUS'd */
