+++ packages/beacon-node/src/network/index.ts
@@ -3,6 +3,6 @@ export * from "./interface.js";
 export * from "./network.js";
 export * from "./nodejs/index.js";
 export * from "./gossip/index.js";
-export * from "./reqresp/index.js";
+export * from "./reqresp/ReqRespBeaconNode.js";
 export * from "./util.js";
 export * from "./peers/index.js";
