+++ packages/beacon-node/test/e2e/network/peers/peerManager.test.ts
@@ -8,7 +8,7 @@ import {BitArray} from "@chainsafe/ssz";
 import {altair, phase0, ssz} from "@lodestar/types";
 import {sleep} from "@lodestar/utils";
 import {createIBeaconConfig} from "@lodestar/config";
-import {IReqRespBeaconNode, ReqRespMethod} from "../../../../src/network/reqresp/index.js";
+import {IReqRespBeaconNode, ReqRespMethod} from "../../../../src/network/reqresp/ReqRespBeaconNode.js";
 import {PeerRpcScoreStore, PeerManager} from "../../../../src/network/peers/index.js";
 import {Eth2Gossipsub, getConnectionsMap, NetworkEvent, NetworkEventBus} from "../../../../src/network/index.js";
 import {PeersData} from "../../../../src/network/peers/peersData.js";
